Move routes to separate directory and isolate pool_manager
