+++ src/odm2_postgres_api/utils/api_pool_manager.py
@@ -1,0 +1,13 @@
+import asyncpg
+
+
+class ApiPoolManager:
+    def __init__(self):
+        self.pool = None  # Optional[asyncpg.pool.Pool]
+
+    async def get_conn(self) -> asyncpg.connection.Connection:
+        async with self.pool.acquire() as connection:
+            yield connection
+
+
+api_pool_manager = ApiPoolManager()
