+++ tests/test_utils.py
@@ -26,3 +26,12 @@ async def truncate_all_data(connection, schema: str):
     rows = await connection.fetch("select * from pg_tables where schemaname=$1;", schema)
     tables = [t["tablename"] for t in rows]
     await _truncate_tables(schema=schema, tables=tables)
+
+
+def test_pool_mamager_uses_same_instance():
+    from odm2_postgres_api.routes.shared import shared_routes
+    from odm2_postgres_api.routes.begroing_routes import begroing_routes
+    from odm2_postgres_api.app import api_pool_manager
+
+    assert shared_routes.api_pool_manager is begroing_routes.api_pool_manager
+    assert api_pool_manager is begroing_routes.api_pool_manager
